api-bots/requirements.txt
discord.py>=2.3.0,<3.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
python-dotenv>=1.0.1
